fixed annoying object placement collision bug

+++ include/World/ChunkPosition.hpp
@@ -2,7 +2,7 @@
 
 struct ChunkPosition
 {
-    int16_t x, y;
+    int x, y;
     ChunkPosition() : x(0), y(0) {}
     ChunkPosition(int _x, int _y) : x(_x), y(_y) {}
     bool operator==(const ChunkPosition& other) const
